Files added for review by Jonathan
Jonathan - this is the set of instructions (in the .pdf file) along with the other files in my netbeans folder for the first 3 of 4 coding projects in assignment 1.  I believe I have completed them successfully.  Maybe you can review and provide input.  I hope to have # 4 added, possibly in the morning before you have a chance to review the first 3.  The set of 4 is due on Tuesday evening at 1159PM.